Issue #13: Check for file before opening application.xml
The script now checks to make sure the file can be read before
comparing the versions. Further, it will fail more gracefully
if the file is not found, along with giving a brief explanation.

Ideally, I'd like to give users a chance to interract with the
script here, and give them the option of (re)moving the file
for them at this point.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>